Add access controls to trip resource. Test failing

+++ resource/resource_trip.go
@@ -8,6 +8,7 @@ import (
 	"github.com/LewisWatson/carshare-back/storage"
 	"github.com/benbjohnson/clock"
 	"github.com/manyminds/api2go"
+	"gopkg.in/LewisWatson/firebase-jwt-auth.v1"
 )
 
 // TripResource for api2go routes
@@ -15,39 +16,31 @@ type TripResource struct {
 	TripStorage     storage.TripStorage
 	UserStorage     storage.UserStorage
 	CarShareStorage storage.CarShareStorage
+	TokenVerifier   fireauth.TokenVerifier
 	Clock           clock.Clock
 }
 
 // FindAll to satisfy api2go.FindAll interface
 func (t TripResource) FindAll(r api2go.Request) (api2go.Responder, error) {
-	trips, err := t.TripStorage.GetAll(r.Context)
-	if err != nil {
-		return &Response{}, api2go.NewHTTPError(
-			fmt.Errorf("Error retrieveing all trips, %s", err),
-			http.StatusText(http.StatusInternalServerError),
-			http.StatusInternalServerError,
-		)
-	}
-	/*
-	 * Populate the trip relationships. If an error occurs then return the error
-	 * along with what has been retrieved up to that point
-	 */
-	for _, trip := range trips {
-		err = t.populate(&trip, r.Context)
-		if err != nil {
-			errMsg := fmt.Sprintf("Error when populating trip %s", trip.GetID())
-			return &Response{Res: trips}, api2go.NewHTTPError(
-				fmt.Errorf("%s, %s", errMsg, err),
-				errMsg,
-				http.StatusInternalServerError,
-			)
-		}
-	}
-	return &Response{Res: trips}, nil
+	return &Response{}, api2go.NewHTTPError(
+		fmt.Errorf("Find all trips not supported"),
+		http.StatusText(http.StatusMethodNotAllowed),
+		http.StatusMethodNotAllowed,
+	)
 }
 
 // FindOne to satisfy api2go.CRUD interface
 func (t TripResource) FindOne(ID string, r api2go.Request) (api2go.Responder, error) {
+
+	requestingUser, err := getRequestUser(r, t.TokenVerifier, t.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			err,
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	trip, err := t.TripStorage.GetOne(ID, r.Context)
 	switch err {
 	case nil:
@@ -66,6 +59,29 @@ func (t TripResource) FindOne(ID string, r api2go.Request) (api2go.Responder, er
 			http.StatusInternalServerError,
 		)
 	}
+
+	httpErr := t.checkUserIsMemberOfCarShare(requestingUser, trip, r.Context)
+	if httpErr.Errors == nil {
+		return &Response{}, httpErr
+	}
+
+	carShare, err := t.CarShareStorage.GetOne(trip.CarShareID, r.Context)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("unable to find car share %s linked to trip %s: %s", trip.CarShareID, trip.GetID(), err),
+			"Error finding associated car share",
+			http.StatusInternalServerError,
+		)
+	}
+
+	if !carShare.IsMember(requestingUser.GetID()) && !carShare.IsAdmin(requestingUser.GetID()) {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("user %s attempting to access trip %s for which they are not a member or admin for", requestingUser.GetID(), trip.GetID()),
+			"must be a member or admin for associated carshare to access",
+			http.StatusForbidden,
+		)
+	}
+
 	// if an error occurs while populating, still attempt to send the remainder
 	// of the response
 	popErr := t.populate(&trip, r.Context)
@@ -77,11 +93,22 @@ func (t TripResource) FindOne(ID string, r api2go.Request) (api2go.Responder, er
 			http.StatusInternalServerError,
 		)
 	}
+
 	return &Response{Res: trip}, err
 }
 
 // Create to satisfy api2go.CRUD interface
 func (t TripResource) Create(obj interface{}, r api2go.Request) (api2go.Responder, error) {
+
+	requestingUser, err := getRequestUser(r, t.TokenVerifier, t.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			err,
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	trip, ok := obj.(model.Trip)
 	if !ok {
 		return &Response{}, api2go.NewHTTPError(
@@ -89,6 +116,11 @@ func (t TripResource) Create(obj interface{}, r api2go.Request) (api2go.Responde
 			http.StatusText(http.StatusBadRequest),
 			http.StatusBadRequest,
 		)
+	}
+
+	httpErr := t.checkUserIsMemberOfCarShare(requestingUser, trip, r.Context)
+	if httpErr.Errors == nil {
+		return &Response{}, httpErr
 	}
 
 	trip.Scores = make(map[string]model.Score)
@@ -144,6 +176,15 @@ func (t TripResource) Create(obj interface{}, r api2go.Request) (api2go.Responde
 // Delete to satisfy the api2go.CRUD interface
 func (t TripResource) Delete(id string, r api2go.Request) (api2go.Responder, error) {
 
+	requestingUser, err := getRequestUser(r, t.TokenVerifier, t.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			err,
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	trip, err := t.TripStorage.GetOne(id, r.Context)
 	switch err {
 	case nil:
@@ -161,6 +202,11 @@ func (t TripResource) Delete(id string, r api2go.Request) (api2go.Responder, err
 			errMsg,
 			http.StatusInternalServerError,
 		)
+	}
+
+	httpErr := t.checkUserIsMemberOfCarShare(requestingUser, trip, r.Context)
+	if httpErr.Errors == nil {
+		return &Response{}, httpErr
 	}
 
 	err = t.TripStorage.Delete(id, r.Context)
@@ -208,6 +254,16 @@ func (t TripResource) Delete(id string, r api2go.Request) (api2go.Responder, err
 
 // Update to satisfy api2go.CRUD interface
 func (t TripResource) Update(obj interface{}, r api2go.Request) (api2go.Responder, error) {
+
+	requestingUser, err := getRequestUser(r, t.TokenVerifier, t.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			err,
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	trip, ok := obj.(model.Trip)
 	if !ok {
 		return &Response{}, api2go.NewHTTPError(
@@ -216,72 +272,53 @@ func (t TripResource) Update(obj interface{}, r api2go.Request) (api2go.Responde
 			http.StatusBadRequest,
 		)
 	}
-	if trip.CarShareID != "" {
-		/*
-		 * Prevent trips from being changing car shares after they are initially assigned
-		 */
-		tripInDataStore, err := t.TripStorage.GetOne(trip.GetID(), r.Context)
-		if err != nil {
-			if err == storage.ErrNotFound {
-				return &Response{}, api2go.NewHTTPError(
-					fmt.Errorf("Unable to find trip %s to update", trip.GetID()),
-					http.StatusText(http.StatusNotFound),
-					http.StatusNotFound,
-				)
-			}
-			errMsg := fmt.Sprintf("Error retrieving trip %s", trip.GetID())
-			return &Response{}, api2go.NewHTTPError(
-				fmt.Errorf("%s, %s", errMsg, err),
-				errMsg,
-				http.StatusInternalServerError,
-			)
-		}
-		if tripInDataStore.CarShareID != "" && tripInDataStore.CarShareID != trip.CarShareID {
-			errMsg := fmt.Sprintf("trip %s already belongs to another car share", trip.GetID())
-			return &Response{}, api2go.NewHTTPError(
-				fmt.Errorf("%s", errMsg),
-				errMsg,
-				http.StatusInternalServerError,
-			)
-		}
 
-		// add trip to car shares trip list
-		carShare, err := t.CarShareStorage.GetOne(trip.CarShareID, r.Context)
-		if err != nil {
-			if err == storage.ErrNotFound {
-				err = fmt.Errorf("Unable to find car share %s to in order to add trip relationship", trip.CarShareID)
-				return &Response{}, api2go.NewHTTPError(
-					err,
-					err.Error(),
-					http.StatusInternalServerError,
-				)
-			}
-			errMsg := fmt.Sprintf("Error retrieving car share %s in order to add trip relationship", trip.CarShareID)
-			return &Response{}, api2go.NewHTTPError(
-				fmt.Errorf("%s, %s", errMsg, err),
-				errMsg,
-				http.StatusInternalServerError,
-			)
-		}
-		found := false
-		for _, tripID := range carShare.TripIDs {
-			if tripID == trip.GetID() {
-				found = true
-				break
-			}
-		}
-		if !found {
-			carShare.TripIDs = append(carShare.TripIDs, trip.GetID())
-			err = t.CarShareStorage.Update(carShare, r.Context)
-			if err != nil {
-				errMsg := fmt.Sprintf("Error updating car share %s", carShare.GetID())
-				return &Response{}, api2go.NewHTTPError(
-					fmt.Errorf("%s, %s", errMsg, err),
-					errMsg,
-					http.StatusInternalServerError,
-				)
-			}
-		}
+	if trip.CarShareID == "" {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Invalid instance given to trip update (missing carShareID): %v", obj),
+			"must provide a carShareID",
+			http.StatusBadRequest,
+		)
+	}
+
+	tripInDataStore, err := t.TripStorage.GetOne(trip.GetID(), r.Context)
+	switch err {
+	case nil:
+		break
+	case storage.ErrNotFound:
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Unable to find trip %s to update", trip.GetID()),
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
+		)
+	default:
+		errMsg := fmt.Sprintf("Error retrieving trip %s", trip.GetID())
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("%s, %s", errMsg, err),
+			errMsg,
+			http.StatusInternalServerError,
+		)
+	}
+
+	// important to check against the trip in the data store
+	httpErr := t.checkUserIsMemberOfCarShare(requestingUser, tripInDataStore, r.Context)
+	if httpErr.Errors == nil {
+		return &Response{}, httpErr
+	}
+
+	// Prevent trips from being re-assigned car shares
+	if tripInDataStore.CarShareID != "" && tripInDataStore.CarShareID != trip.CarShareID {
+		errMsg := fmt.Sprintf("trip %s already belongs to another car share", trip.GetID())
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("%s", errMsg),
+			errMsg,
+			http.StatusInternalServerError,
+		)
+	}
+
+	httpErr = t.addToCarShareTripList(trip, r.Context)
+	if httpErr.Errors != nil {
+		return &Response{}, httpErr
 	}
 
 	// verify driver
@@ -331,7 +368,6 @@ func (t TripResource) Update(obj interface{}, r api2go.Request) (api2go.Responde
 	trip.CalculateScores(latestTrip.Scores)
 
 	err = t.TripStorage.Update(trip, r.Context)
-
 	switch err {
 	case nil:
 		break
@@ -365,6 +401,54 @@ func (t TripResource) Update(obj interface{}, r api2go.Request) (api2go.Responde
 	return &Response{Res: trip, Code: http.StatusNoContent}, err
 }
 
+// addToCarShareTripList updates the associated carshare and ensures that it has the trip in its list of trips
+func (t TripResource) addToCarShareTripList(trip model.Trip, ctx api2go.APIContexter) api2go.HTTPError {
+
+	carShare, err := t.CarShareStorage.GetOne(trip.CarShareID, ctx)
+	switch err {
+	case nil:
+		break
+	case storage.ErrNotFound:
+		err = fmt.Errorf("Unable to find car share %s to in order to add trip relationship", trip.CarShareID)
+		return api2go.NewHTTPError(
+			err,
+			err.Error(),
+			http.StatusInternalServerError,
+		)
+	default:
+		errMsg := fmt.Sprintf("Error retrieving car share %s in order to add trip relationship", trip.CarShareID)
+		return api2go.NewHTTPError(
+			fmt.Errorf("%s, %s", errMsg, err),
+			errMsg,
+			http.StatusInternalServerError,
+		)
+	}
+
+	found := false
+
+	for _, tripID := range carShare.TripIDs {
+		if tripID == trip.GetID() {
+			found = true
+			break
+		}
+	}
+
+	if !found {
+		carShare.TripIDs = append(carShare.TripIDs, trip.GetID())
+		err = t.CarShareStorage.Update(carShare, ctx)
+		if err != nil {
+			errMsg := fmt.Sprintf("Error updating car share %s", carShare.GetID())
+			return api2go.NewHTTPError(
+				fmt.Errorf("%s, %s", errMsg, err),
+				errMsg,
+				http.StatusInternalServerError,
+			)
+		}
+	}
+
+	return api2go.HTTPError{}
+}
+
 // populate the relationships for a trip
 func (t TripResource) populate(trip *model.Trip, context api2go.APIContexter) error {
 
@@ -387,4 +471,26 @@ func (t TripResource) populate(trip *model.Trip, context api2go.APIContexter) er
 	}
 
 	return nil
+}
+
+func (t TripResource) checkUserIsMemberOfCarShare(user model.User, trip model.Trip, ctx api2go.APIContexter) api2go.HTTPError {
+
+	carShare, err := t.CarShareStorage.GetOne(trip.CarShareID, ctx)
+	if err != nil {
+		return api2go.NewHTTPError(
+			fmt.Errorf("unable to find car share %s linked to trip %s: %s", trip.CarShareID, trip.GetID(), err),
+			"Error finding associated car share",
+			http.StatusInternalServerError,
+		)
+	}
+
+	if !carShare.IsMember(user.GetID()) {
+		return api2go.NewHTTPError(
+			fmt.Errorf("user %s attempting to access trip %s for car share %s they are not a member of", user.GetID(), trip.GetID(), trip.CarShareID),
+			"must be a member or admin for associated carshare to access",
+			http.StatusForbidden,
+		)
+	}
+
+	return api2go.HTTPError{}
 }
